refactor(ckpt): isolate direct FP8 export helpers
